+++ lib/widgets/navbar.dart
@@ -1,11 +1,13 @@
 import 'package:every_door/helpers/tile_layers.dart';
 import 'package:every_door/providers/api_status.dart';
 import 'package:every_door/providers/changes.dart';
+import 'package:every_door/providers/changeset_tags.dart';
 import 'package:every_door/providers/editor_mode.dart';
 import 'package:every_door/providers/editor_settings.dart';
 import 'package:every_door/providers/imagery.dart';
 import 'package:every_door/providers/notes.dart';
 import 'package:every_door/providers/uploader.dart';
+import 'package:every_door/screens/settings/changeset_pane.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -16,6 +18,15 @@ class BrowserNavigationBar extends ConsumerWidget {
   const BrowserNavigationBar({Key? key, required this.downloadAmenities})
       : super(key: key);
 
+  Future<bool> _showChangesetPane(BuildContext context) async {
+    final result = await showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      builder: (context) => ChangesetSheetPane(),
+    );
+    return result != false;
+  }
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final editorMode = ref.watch(editorModeProvider);
@@ -23,6 +34,8 @@ class BrowserNavigationBar extends ConsumerWidget {
     final hasChangesToUpload = ref.watch(changesProvider).haveNoErrorChanges();
     final hasNotesToUpload = ref.watch(notesProvider).haveChanges;
     final double bottomPadding = MediaQuery.of(context).padding.bottom;
+    final bool haveHashtags =
+        ref.watch(changesetTagsProvider).getHashtags().isNotEmpty;
     final loc = AppLocalizations.of(context)!;
 
     IconButton dataButton;
@@ -43,9 +56,29 @@ class BrowserNavigationBar extends ConsumerWidget {
         onPressed: apiStatus != ApiStatus.idle
             ? null
             : () async {
-                ref.read(uploaderProvider).upload(context);
+                final review = ref.read(editorSettingsProvider).changesetReview;
+                bool needAsk = review == ChangesetReview.always ||
+                    (haveHashtags && review == ChangesetReview.withTags);
+                bool upload = true;
+                if (needAsk) {
+                  if (!await _showChangesetPane(context)) upload = false;
+                }
+                // We won't lose the context here.
+                // ignore: use_build_context_synchronously
+                if (upload) ref.read(uploaderProvider).upload(context);
               },
-        icon: Icon(Icons.upload),
+        icon: Stack(children: [
+          Icon(Icons.upload),
+          if (haveHashtags)
+            Positioned(
+              child: Text(
+                '#',
+                style: TextStyle(color: Colors.yellow, fontSize: 10.0),
+              ),
+              right: 0.0,
+              top: 0.0,
+            ),
+        ]),
         tooltip: loc.navUpload,
         color: Colors.yellow,
         disabledColor: Colors.yellow.withOpacity(0.2),